refactor(auth, profile): update API tags and DTO paths
- Changed API tag from 'auth' to 'authentication' in `AuthController`.
- Moved profile DTOs to a common `dto/profile` directory.
- Updated imports in `ProfileController` and `ProfileService` to reflect new DTO paths.
- Removed unused `Profile` entity.
