--- src/profile/entities/profile.entity.ts
@@ -1,1 +1,0 @@
-export class Profile {}
