+++ src/profile/profile.controller.ts
@@ -1,7 +1,7 @@
 import { Controller, Get, Post, Body, Patch, Param, Delete } from '@nestjs/common';
 import { ProfileService } from './profile.service';
-import { CreateProfileDto } from './dto/create-profile.dto';
-import { UpdateProfileDto } from './dto/update-profile.dto';
+import { CreateProfileDto } from '../dto/profile/create-profile.dto';
+import { UpdateProfileDto } from '../dto/profile/update-profile.dto';
 
 @Controller('profile')
 export class ProfileController {
