+++ src/profile/profile.service.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@nestjs/common';
-import { CreateProfileDto } from './dto/create-profile.dto';
-import { UpdateProfileDto } from './dto/update-profile.dto';
+import { CreateProfileDto } from '../dto/profile/create-profile.dto';
+import { UpdateProfileDto } from '../dto/profile/update-profile.dto';
 
 @Injectable()
 export class ProfileService {
